feat(platform): minor changes
