+++ apps/platform/app/page.tsx
@@ -1,4 +1,4 @@
-import Operations from '@components/landing/operations'
+import { Operations } from '@components/landing/operations'
 
 import { CTA } from '@components/landing/cta'
 import { Boxes } from '@components/landing/boxes'
@@ -10,6 +10,7 @@ export default function Home(): React.ReactNode {
   return (
     <main className="bg-background min-h-screen pb-10">
       <Navbar />
+      <Operations />
       <CTA />
       <div className="h-[50dvh]"></div>
       <Features />
